create dir EditorsToolbars, moved Left+RightEditorToolbar, create component EditorHeightButtons
