+++ services/app/assets/js/widgets/containers/EditorsToolbars/RightEditorToolbar.jsx
@@ -1,31 +1,12 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import _ from 'lodash';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import * as selectors from '../selectors';
-import { compressEditorHeight, expandEditorHeight } from '../middlewares/Game';
-import LanguagePicker from '../components/LanguagePicker';
-import UserInfo from './UserInfo';
-import GameResultIcon from '../components/GameResultIcon';
-
-const renderEditorHeightButtons = (compressEditor, expandEditor, userId) => (
-  <div className="btn-group btn-group-sm mr-2" role="group" aria-label="Editor height">
-    <button
-      type="button"
-      className="btn btn-sm btn-light border rounded"
-      onClick={compressEditor(userId)}
-    >
-      <i className="fas fa-compress-arrows-alt" aria-hidden="true" />
-    </button>
-    <button
-      type="button"
-      className="btn btn-sm btn-light border rounded ml-2"
-      onClick={expandEditor(userId)}
-    >
-      <i className="fas fa-expand-arrows-alt" aria-hidden="true" />
-    </button>
-  </div>
-);
+import * as selectors from '../../selectors';
+import LanguagePicker from '../../components/LanguagePicker';
+import UserInfo from '../UserInfo';
+import GameResultIcon from '../../components/GameResultIcon';
+import EditorHeightButtons from './EditorHeightButtons';
 
 const renderNameplate = (player = {}, onlineUsers) => {
   const isOnline = _.find(onlineUsers, { id: player.id });
@@ -52,10 +33,6 @@ const RightEditorToolbar = () => {
   const rightEditorLangSlug = useSelector(state => selectors.userLangSelector(rightUserId)(state));
   const players = useSelector(state => selectors.gamePlayersSelector(state));
 
-  const dispatch = useDispatch();
-  const compressEditor = userId => () => dispatch(compressEditorHeight(userId));
-  const expandEditor = userId => () => dispatch(expandEditorHeight(userId));
-
   if (rightEditorLangSlug === null) { return null; }
 
   return (
@@ -70,7 +47,9 @@ const RightEditorToolbar = () => {
       />
       {renderNameplate(players[rightUserId], onlineUsers)}
       <div className="ml-auto btn-group" role="group" aria-label="Editor settings">
-        {renderEditorHeightButtons(compressEditor, expandEditor, rightUserId)}
+        <EditorHeightButtons
+          typeEditor="right"
+        />
         <LanguagePicker
           languages={languages}
           currentLangSlug={rightEditorLangSlug}
