+++ services/app/assets/js/widgets/containers/GameWidget.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 // import Gon from 'gon';
 import * as selectors from '../selectors';
 import Editor from './Editor';
-import LeftEditorToolbar from './LeftEditorToolbar';
-import RightEditorToolbar from './RightEditorToolbar';
+import LeftEditorToolbar from './EditorsToolbars/LeftEditorToolbar';
+import RightEditorToolbar from './EditorsToolbars/RightEditorToolbar';
 import GameActionButtons from '../components/GameActionButtons';
 import { sendEditorText } from '../middlewares/Game';
 import ExecutionOutput from '../components/ExecutionOutput';
